fix: profiler threshold in C++

+++ src/library/profiling.cpp
@@ -14,7 +14,7 @@ bool get_profiler(options const & opts) {
     return lean_get_profiler(opts.to_obj_arg());
 }
 
-extern "C" float lean_get_profiler_threshold(obj_arg opts);
+extern "C" double lean_get_profiler_threshold(obj_arg opts);
 second_duration get_profiling_threshold(options const & opts) {
     double ms = lean_get_profiler_threshold(opts.to_obj_arg());
     return second_duration(ms);
